update rp functions
